+++ AllOfSoccer/SupporttingFiles/Assets.xcassets/CurrencyCircleDollar.imageset/CurrencyCircleDollar.pdf
@@ -1,0 +1,125 @@
+%PDF-1.7
+
+1 0 obj
+  << >>
+endobj
+
+2 0 obj
+  << /Length 3 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 1.687500 1.545898 cm
+0.200000 0.200000 0.200000 scn
+7.312503 14.766602 m
+5.866227 14.766602 4.452430 14.337730 3.249896 13.534224 c
+2.047361 12.730715 1.110100 11.588659 0.556634 10.252474 c
+0.003168 8.916290 -0.141643 7.445989 0.140511 6.027503 c
+0.422665 4.609017 1.119113 3.306054 2.141785 2.283383 c
+3.164456 1.260712 4.467419 0.564264 5.885905 0.282109 c
+7.304391 -0.000045 8.774691 0.144767 10.110875 0.698233 c
+11.447061 1.251698 12.589117 2.188960 13.392625 3.391495 c
+14.196133 4.594029 14.625003 6.007826 14.625003 7.454102 c
+14.622808 9.392824 13.851683 11.251513 12.480798 12.622396 c
+11.109914 13.993280 9.251225 14.764407 7.312503 14.766602 c
+7.312503 14.766602 l
+h
+8.156253 4.079102 m
+7.875003 4.079102 l
+7.875003 3.516602 l
+7.875003 3.367417 7.815741 3.224342 7.710251 3.118854 c
+7.604762 3.013365 7.461688 2.954102 7.312503 2.954102 c
+7.163319 2.954102 7.020245 3.013365 6.914756 3.118854 c
+6.809267 3.224342 6.750003 3.367417 6.750003 3.516602 c
+6.750003 4.079102 l
+5.625003 4.079102 l
+5.475819 4.079102 5.332745 4.138365 5.227256 4.243854 c
+5.121767 4.349342 5.062503 4.492417 5.062503 4.641602 c
+5.062503 4.790786 5.121767 4.933861 5.227256 5.039350 c
+5.332745 5.144838 5.475819 5.204102 5.625003 5.204102 c
+8.156253 5.204102 l
+8.380030 5.204102 8.594641 5.292996 8.752874 5.451231 c
+8.911108 5.609465 9.000003 5.824075 9.000003 6.047852 c
+9.000003 6.271628 8.911108 6.486238 8.752874 6.644472 c
+8.594641 6.802706 8.380030 6.891602 8.156253 6.891602 c
+6.468753 6.891602 l
+5.946608 6.891602 5.445849 7.099023 5.076637 7.468235 c
+4.707425 7.837448 4.500003 8.338207 4.500003 8.860352 c
+4.500003 9.382496 4.707425 9.883255 5.076637 10.252468 c
+5.445849 10.621680 5.946608 10.829102 6.468753 10.829102 c
+6.750003 10.829102 l
+6.750003 11.391602 l
+6.750003 11.540786 6.809267 11.683860 6.914756 11.789349 c
+7.020245 11.894838 7.163319 11.954102 7.312503 11.954102 c
+7.461688 11.954102 7.604762 11.894838 7.710251 11.789349 c
+7.815741 11.683860 7.875003 11.540786 7.875003 11.391602 c
+7.875003 10.829102 l
+9.000003 10.829102 l
+9.149187 10.829102 9.292262 10.769838 9.397751 10.664349 c
+9.503241 10.558860 9.562503 10.415786 9.562503 10.266602 c
+9.562503 10.117417 9.503241 9.974343 9.397751 9.868855 c
+9.292262 9.763365 9.149187 9.704102 9.000003 9.704102 c
+6.468753 9.704102 l
+6.244977 9.704102 6.030365 9.615207 5.872132 9.456973 c
+5.713898 9.298739 5.625003 9.084127 5.625003 8.860352 c
+5.625003 8.636576 5.713898 8.421964 5.872132 8.263730 c
+6.030365 8.105496 6.244977 8.016602 6.468753 8.016602 c
+8.156253 8.016602 l
+8.678398 8.016602 9.179158 7.809181 9.548370 7.439969 c
+9.917583 7.070756 10.125003 6.569997 10.125003 6.047852 c
+10.125003 5.525706 9.917583 5.024947 9.548370 4.655735 c
+9.179158 4.286523 8.678398 4.079102 8.156253 4.079102 c
+8.156253 4.079102 l
+h
+f
+n
+Q
+
+endstream
+endobj
+
+3 0 obj
+  2857
+endobj
+
+4 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 18.000000 18.000000 ]
+     /Resources 1 0 R
+     /Contents 2 0 R
+     /Parent 5 0 R
+  >>
+endobj
+
+5 0 obj
+  << /Kids [ 4 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+6 0 obj
+  << /Type /Catalog
+     /Pages 5 0 R
+  >>
+endobj
+
+xref
+0 7
+0000000000 65535 f
+0000000010 00000 n
+0000000034 00000 n
+0000002947 00000 n
+0000002970 00000 n
+0000003143 00000 n
+0000003217 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 6 0 R
+   /Size 7
+>>
+startxref
+3276
+%%EOF
