+++ AllOfSoccer/SupporttingFiles/Assets.xcassets/UserCircle.imageset/UserCircle.pdf
@@ -1,0 +1,105 @@
+%PDF-1.7
+
+1 0 obj
+  << >>
+endobj
+
+2 0 obj
+  << /Length 3 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 5.906250 6.468750 cm
+0.000000 0.000000 0.000000 scn
+3.093750 0.000000 m
+4.802381 0.000000 6.187500 1.385119 6.187500 3.093750 c
+6.187500 4.802381 4.802381 6.187500 3.093750 6.187500 c
+1.385119 6.187500 0.000000 4.802381 0.000000 3.093750 c
+0.000000 1.385119 1.385119 0.000000 3.093750 0.000000 c
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 1.687500 1.545898 cm
+0.000000 0.000000 0.000000 scn
+7.312503 14.766602 m
+5.866227 14.766602 4.452430 14.337730 3.249896 13.534224 c
+2.047361 12.730715 1.110100 11.588659 0.556634 10.252474 c
+0.003168 8.916290 -0.141643 7.445989 0.140511 6.027503 c
+0.422665 4.609017 1.119113 3.306054 2.141785 2.283383 c
+3.164456 1.260712 4.467419 0.564264 5.885905 0.282109 c
+7.304391 -0.000045 8.774691 0.144767 10.110875 0.698233 c
+11.447061 1.251698 12.589117 2.188960 13.392625 3.391495 c
+14.196133 4.594029 14.625003 6.007826 14.625003 7.454102 c
+14.622808 9.392824 13.851683 11.251513 12.480798 12.622396 c
+11.109914 13.993280 9.251225 14.764407 7.312503 14.766602 c
+7.312503 14.766602 l
+h
+11.936116 3.347137 m
+11.484566 4.001341 10.899005 4.551996 10.218318 4.962540 c
+9.435587 4.214987 8.394863 3.797852 7.312503 3.797852 c
+6.230143 3.797852 5.189420 4.214987 4.406688 4.962540 c
+3.725998 4.551994 3.140432 4.001340 2.688873 3.347137 c
+1.896105 4.238573 1.377934 5.340176 1.196758 6.519292 c
+1.015582 7.698407 1.179127 8.904758 1.667698 9.993076 c
+2.156268 11.081394 2.949033 12.005275 3.950527 12.653463 c
+4.952021 13.301651 6.119541 13.646509 7.312495 13.646509 c
+8.505448 13.646509 9.672968 13.301651 10.674462 12.653463 c
+11.675957 12.005275 12.468721 11.081394 12.957293 9.993076 c
+13.445864 8.904758 13.609407 7.698407 13.428231 6.519292 c
+13.247056 5.340176 12.728885 4.238573 11.936116 3.347137 c
+11.936116 3.347137 l
+h
+f
+n
+Q
+
+endstream
+endobj
+
+3 0 obj
+  1836
+endobj
+
+4 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 18.000000 18.000000 ]
+     /Resources 1 0 R
+     /Contents 2 0 R
+     /Parent 5 0 R
+  >>
+endobj
+
+5 0 obj
+  << /Kids [ 4 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+6 0 obj
+  << /Type /Catalog
+     /Pages 5 0 R
+  >>
+endobj
+
+xref
+0 7
+0000000000 65535 f
+0000000010 00000 n
+0000000034 00000 n
+0000001926 00000 n
+0000001949 00000 n
+0000002122 00000 n
+0000002196 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 6 0 R
+   /Size 7
+>>
+startxref
+2255
+%%EOF
